commit for WSL testing

+++ change-text-in-file/changeTextInFile.py
@@ -7,28 +7,70 @@ print('Running changeTextInFile.py (Use CTRL + C to exit)')
 
 
 # Variables:
+isDirectory = ''
 
-startingPath = pathlib.Path.home()
-filePath = input(f'File Name (starting at {startingPath}): ') # set file path here
-finalFilePath = path.join(startingPath, filePath)
-print (f'Selected filePath: {finalFilePath}')
+while (isDirectory != 'y' and isDirectory != 'n'):
+  isDirectory = input('Are you trying to change multiple files in a directory? (y/n) ')
 
-currentText = input('Current Text (the text you want replaced): ') # set text you want to change here
-print(f'Current text set to: {currentText}')
+  if (isDirectory != 'y' and isDirectory != 'n'):
+    print('Invalid input. Please try again.')
 
-replacementText = input('Replacement Text: ') # set replacement text here
-print(f'Replacement text set to: {replacementText}')
+if (isDirectory == 'n'):
 
+  startingPath = pathlib.Path.home()
+  filePath = input(f'File Name (starting at {startingPath}): ') # set file path here
+  finalFilePath = path.join(startingPath, filePath)
+  print (f'Selected filePath: {finalFilePath}')
 
-# Code:
-print('Running...')
+  currentText = input('Current Text (the text you want replaced): ') # set text you want to change here
+  print(f'Current text set to: {currentText}')
 
-with open(finalFilePath, 'r') as file:
-  filedata = file.read()
+  replacementText = input('Replacement Text: ') # set replacement text here
+  print(f'Replacement text set to: {replacementText}')
 
-filedata = filedata.replace(currentText, replacementText)
+  print('Running...')
 
-with open(finalFilePath, 'w') as file:
-  file.write(filedata)
+  with open(finalFilePath, 'r') as file:
+    filedata = file.read()
 
-print('Done!')
+  filedata = filedata.replace(currentText, replacementText)
+
+  with open(finalFilePath, 'w') as file:
+    file.write(filedata)
+
+  print('Done modifying file!')
+
+elif (isDirectory == 'y'):
+
+  startingPath = pathlib.Path.home()
+  directory = input(f'Directory path (starting at {startingPath}): ') # set file path here
+  finalDirectory = path.join(startingPath, directory)
+  print (f'Selected directory: {finalDirectory}')
+
+  currentText = input('Current Text (the text you want replaced): ') # set text you want to change here
+  print(f'Current text set to: {currentText}')
+
+  replacementText = input('Replacement Text: ') # set replacement text here
+  print(f'Replacement text set to: {replacementText}')
+
+  # Code:
+  print('Running...')
+
+  for filePath in pathlib.Path(finalDirectory).iterdir():
+    with open(filePath, 'r') as file:
+      filedata = file.read()
+
+    filedata = filedata.replace(currentText, replacementText)
+
+    with open(filePath, 'w') as file:
+      file.write(filedata)
+
+  # with open(finalFilePath, 'r') as file:
+  #   filedata = file.read()
+
+  # filedata = filedata.replace(currentText, replacementText)
+
+  # with open(finalFilePath, 'w') as file:
+  #   file.write(filedata)
+
+  print('Done modifying files in directory!')
